network: move PokemonCardView to a separate file
